requirements.txt
opacus==1.1.3
timm
pandas
torch==2.3.1+cu121
torchvision==0.18.1+cu121
Pillow
scipy
prv-accountant
transformers
tensorboardX
kornia
diffprivlib